Add localizations

+++ src/api/IReputationService.cs
@@ -1,15 +1,16 @@
 ﻿namespace EloReputation.api;
 
 public interface IReputationService {
-  Task<float> GetReputation(ulong steamId);
+  Task<double> GetReputation(ulong steamId);
 
-  Task<IEnumerable<(ulong, float)>> GetReputation(IEnumerable<ulong> steamIds) {
+  Task<IEnumerable<(ulong, double)>>
+    GetReputation(IEnumerable<ulong> steamIds) {
     return GetReputation(steamIds.ToArray());
   }
 
-  Task<IEnumerable<(ulong, float)>> GetReputation(params ulong[] steamIds);
+  Task<IEnumerable<(ulong, double)>> GetReputation(params ulong[] steamIds);
 
-  Task<IEnumerable<(ulong, float)>> GetTopReputation(int limit = 10,
+  Task<IEnumerable<(ulong, double)>> GetTopReputation(int limit = 10,
     int offset = 0);
 
   Task AddReputation(ulong giver, ulong receiver, bool positive = true);
